Fixed the vid/pid of the pokken controller
Hopefully added support for an other PS2 controller.

+++ cp_retain_vars.c
@@ -5,8 +5,8 @@
 u8 config_done __attribute__((section(".data"))) = 0;
 
 u8 gHIDSetupDone __attribute__((section(".data"))) = 0;
-u8 gHIDAttached __attribute__((section(".data"))) = 0;
-u8 gHIDCurrentDevice __attribute__((section(".data"))) = 0;
+u16 gHIDAttached __attribute__((section(".data"))) = 0;
+u16 gHIDCurrentDevice __attribute__((section(".data"))) = 0;
 
 HIDClient gHIDClient __attribute__((section(".data")));
 
@@ -17,5 +17,5 @@ HID_Mouse gHID_Mouse __attribute__((section(".data")));
 u8 gHID_Mouse_Mode __attribute__((section(".data"))) = HID_MOUSE_MODE_TOUCH;
 
 u8 config_controller[CONTRPD_MAX_VALUE][CONTRPS_MAX_VALUE][2] __attribute__((section(".data")));
-s8 config_controller_list[CONTRPD_MAX_VALUE] __attribute__((section(".data")));
+u16 config_controller_list[CONTRPD_MAX_VALUE] __attribute__((section(".data")));
 u32 config_controller_data_ptr[CONTRPD_MAX_VALUE][4] __attribute__((section(".data")));
