+++ cp_retain_vars.h
@@ -7,8 +7,8 @@
 extern u8 config_done;
 
 extern u8 gHIDSetupDone;
-extern u8 gHIDAttached;
-extern u8 gHIDCurrentDevice;
+extern u16 gHIDAttached;
+extern u16 gHIDCurrentDevice;
 extern HIDClient gHIDClient;
 
 extern HID_DEVICE_DATA gHID_Devices[CONTRPD_MAX_VALUE];
@@ -18,7 +18,7 @@ extern HID_Mouse gHID_Mouse;
 extern u8 gHID_Mouse_Mode;
 
 extern u8 config_controller[CONTRPD_MAX_VALUE][CONTRPS_MAX_VALUE][2];
-extern s8 config_controller_list[CONTRPD_MAX_VALUE];
+extern u16 config_controller_list[CONTRPD_MAX_VALUE];
 extern u32 config_controller_data_ptr[CONTRPD_MAX_VALUE][4]; //currently max per device
 
 #endif // CP_RETAINS_VARS_H_
